% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{build_dendrogram}
\alias{build_dendrogram}
\title{Convert a nested list to a dendrogram with labels}
\usage{
build_dendrogram(unit_set)
}
\arguments{
\item{unit_set}{A data frame, typically the \code{unit_set} from a \code{build_hierarchy} result.}
}
\value{
An object of class 'dendrogram'
}
\description{
Converts each hierarchy (name) of a list into branch labels of a dendrogram.
}
