% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_map.R
\name{flowbca_map}
\alias{flowbca_map}
\title{Create a Static Map of the Final Clusters}
\usage{
flowbca_map(
  unit_set,
  unit_gis,
  join_col = "sourceunit",
  file_nm = NULL,
  width = 1000
)
}
\arguments{
\item{unit_set}{A data frame from the \code{flowbca} result. Must contain \code{h_parent} and \code{h_level} from \code{build_hierarchy()}.}

\item{unit_gis}{An \code{sf} object with polygons corresponding to the units.}

\item{join_col}{Specifies the join key columns.}

\item{file_nm}{The file name for the output PNG. The default value is set to
the name of unit_set, but the user can specify it (e.g., flowbca_map.png).}

\item{width}{The width of the output image in pixels.}
}
\value{
A list of \code{sf} objects for the clusters and core units.
}
\description{
Generates a static map of the final clusters from a \code{flowbca} analysis.
}
